Fix up tentacle service creds
We weren't actually setting the credentials, and the tests were not working correctly, so it wasn't detecting the issue.

I did try and use the `Set-ServiceCredential` and `Get-ServiceUsername` functions, but ran into issues where the account didn't have "Logon as a service" rights. This is handled inside the `tentacle.exe service` command, so just used that.

Fixes #84