Updated rule check
If the firewall rule did not exist, the Invoke-AndAssert function call was throwing an error because the exit code was 1.  Changed from Invoke-AndAssert to Invoke-Command, capturing the output of the .exe call.  Updated if statement to check for specific text returned by netsh.exe when a rule does not exist.