Added check for firewall rule (#158)
* Added check for firewall rule

I ran into a situation where we ran the DSC multiple times and it added a firewall rule each time it ran.  Added a check to see if the firewall rule already exists

* Changed check to look for DisplayName

Found that Name is a GUID and DisplayName is what we're looking for.

* Updated rule check

Changed check for existing firewall rule to use non-Windows specific methods.

* Updated rule check

If the firewall rule did not exist, the Invoke-AndAssert function call was throwing an error because the exit code was 1.  Changed from Invoke-AndAssert to Invoke-Command, capturing the output of the .exe call.  Updated if statement to check for specific text returned by netsh.exe when a rule does not exist.