Add support for Worker pools to cTentacleAgent + improved drift detection (#171)
* Added check for firewall rule

I ran into a situation where we ran the DSC multiple times and it added a firewall rule each time it ran.  Added a check to see if the firewall rule already exists

* Changed check to look for DisplayName

Found that Name is a GUID and DisplayName is what we're looking for.

* Updated rule check

Changed check for existing firewall rule to use non-Windows specific methods.

* Updated rule check

If the firewall rule did not exist, the Invoke-AndAssert function call was throwing an error because the exit code was 1.  Changed from Invoke-AndAssert to Invoke-Command, capturing the output of the .exe call.  Updated if statement to check for specific text returned by netsh.exe when a rule does not exist.

* Adding worker pools and better drift detection

Added the ability to set worker pools and better drift detection

* Still working

Moving project to another machine

* Fixing drift detection

* Added drift detection and worker pools

* Fixed missing parameters for ports

I missed some variables when I refactored the code which resulted in ports being set incorrectly.  I have fixed the function call of Register-Tentacle in New-Tentacle function as well as the Set-TargetResource when the tentacle is already installed.

* Fixed trailing white spaces and updated tests to work with new code.

* Fixed failing test

Tests were failing due to $OctopusUrl being empty.  Added code to check for that condition.