WIP- Add DHCP publicHostNameConfiguration.
This detects an IP address that's on the host based on a CIDR range, passed in as the CidrRange parameter.

Discovering the IP this way allows for deployments to non-domain-joined
  deployment targets to not require public IP addresses.

This commit does not check that the DHCP lease has expired during Test-*.
 - This is a requirement because a server could be snapshotted and scaled, yet never fix its own tentacle address

I'm unsure on tests at this time, there isn't documentation on how to run the tests.