Convert tests from rspec to pester (#290)
We've had a lot of trouble fighting with ruby - mainly around rubygems. It's just not fun.
Combine that with unfamiliarity and a "why are we using ruby for tests when the rest of the code is in powershell" disconnect, it makes a lot of sense to use Pester as our test framework instead.